requirements.txt
numpy==1.26.3
tensorflow==2.18.0
streamlit==1.30.0
google-generativeai
opencv-python
opencv-python-headless
gdown

